# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrailTeam team management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pods:
    get:
      operationId: listPods
      tags: [pods]
      summary: List all pods
      responses:
        "200":
          description: List of pods with member counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PodWithStats"
    post:
      operationId: createPod
      tags: [pods]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePodBody"
      responses:
        "201":
          description: Created pod
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pod"
  /pods/{id}:
    get:
      operationId: getPod
      tags: [pods]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pod with members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PodWithMembers"
    patch:
      operationId: updatePod
      tags: [pods]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePodBody"
      responses:
        "200":
          description: Updated pod
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pod"
    delete:
      operationId: deletePod
      tags: [pods]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Pod deleted
  /pods/reorder:
    post:
      operationId: reorderPods
      tags: [pods]
      summary: Set pod display order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderPodsBody"
      responses:
        "204":
          description: Reordered successfully

  # ─── EVENTS ──────────────────────────────────────────────────────────────────
...
  /messages:
    get:
      operationId: listBroadcasts
      tags: [messages]
      summary: List broadcasts (messages received/sent)
      parameters:
        - name: podId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of broadcasts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BroadcastWithSender"
    post:
      operationId: sendBroadcast
      tags: [messages]
      summary: Send a broadcast to pods or whole team (coach/admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendBroadcastBody"
      responses:
        "201":
          description: Broadcast sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broadcast"
  /messages/{id}/archive:
    post:
      operationId: archiveBroadcast
      tags: [messages]
      summary: Archive a broadcast (coach/admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Broadcast archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broadcast"
        "404":
          description: Broadcast not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /messages/{id}/unarchive:
    post:
      operationId: unarchiveBroadcast
      tags: [messages]
      summary: Restore an archived broadcast (coach/admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Broadcast restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broadcast"
        "404":
          description: Broadcast not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /messages/contact-coach:
    post:
      operationId: contactCoach
      tags: [messages]
      summary: Parent sends a message to a specific coach or all pod coaches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactCoachBody"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── BOARD ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ─── USER ─────────────────────────────────────────────────────────────────
...
    UserNotificationPreferences:
      type: object
      properties:
        practiceReminders:
          type: boolean
        coachMessages:
          type: boolean
        carpoolUpdates:
          type: boolean
        eventReminders:
          type: boolean
        rosterUpdates:
          type: boolean
        boardReplies:
          type: boolean
      required:
        - practiceReminders
        - coachMessages
        - carpoolUpdates
        - eventReminders
        - rosterUpdates
        - boardReplies
    User:
      type: object
      properties:
        id:
          type: integer
        householdId:
          type: integer
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [super_admin, coach, parent, student]
        podId:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        gender:
          type: string
          nullable: true
        grade:
          type: integer
          nullable: true
        coachCertLevel:
          type: string
          nullable: true
        notificationsEnabled:
          type: boolean
        emailNotifications:
          type: boolean
        smsNotifications:
          type: boolean
        pushNotifications:
          type: boolean
        defaultCarpoolSeats:
          type: integer
          nullable: true
        defaultCarpoolTrays:
          type: integer
          nullable: true
        notificationPreferences:
          $ref: '#/components/schemas/UserNotificationPreferences'
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - email
        - role
        - isActive
        - notificationsEnabled
        - emailNotifications
        - smsNotifications
        - pushNotifications
        - createdAt
...
    Pod:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        headCoachId:
          type: integer
          nullable: true
        color:
          type: string
          nullable: true
        season:
          type: string
          nullable: true
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, isActive, sortOrder, createdAt]
    ReorderPodsBody:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
      required: [ids]
    PodWithStats:
      allOf:
        - $ref: "#/components/schemas/Pod"
        - type: object
          properties:
            studentCount:
              type: integer
            coachCount:
              type: integer
            complianceRate:
              type: number
          required: [studentCount, coachCount, complianceRate]
    PodWithMembers:
      allOf:
        - $ref: "#/components/schemas/Pod"
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/User"
            coaches:
              type: array
              items:
                $ref: "#/components/schemas/User"
          required: [members, coaches]
    CreatePodBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        headCoachId:
          type: integer
        color:
          type: string
        season:
          type: string
      required: [name]
    UpdatePodBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        headCoachId:
          type: integer
        color:
          type: string
        season:
          type: string
        isActive:
          type: boolean

    # ─── EVENT ────────────────────────────────────────────────────────────────
...
    Broadcast:
      type: object
      properties:
        id:
          type: integer
        senderUserId:
          type: integer
          nullable: true
        subject:
          type: string
          nullable: true
        body:
          type: string
        channel:
          type: string
          enum: [email, sms, push]
        targetPodIds:
          type: array
          items:
            type: string
          nullable: true
        isAllTeam:
          type: boolean
        recipientCount:
          type: integer
        deliveredCount:
          type: integer
          description: Number of emails successfully delivered (null if email not configured)
          nullable: true
        failedCount:
          type: integer
          description: Number of emails that failed to deliver (null if email not configured)
          nullable: true
        emailConfigured:
          type: boolean
          description: Whether email delivery is configured for this install
        sentAt:
          type: string
          format: date-time
          nullable: true
        archivedAt:
          type: string
          format: date-time
          nullable: true
          description: Set when the broadcast is archived; null when active
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - body
        - channel
        - isAllTeam
        - recipientCount
        - emailConfigured
        - createdAt
    BroadcastWithSender:
      allOf:
        - $ref: "#/components/schemas/Broadcast"
        - type: object
          properties:
            sender:
              $ref: "#/components/schemas/User"
              nullable: true
    SendBroadcastBody:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        channel:
          type: string
          enum: [email, sms, push]
        targetPodIds:
          type: array
          items:
            type: string
        isAllTeam:
          type: boolean
      required: [body, channel]
    ContactCoachBody:
      type: object
      properties:
        coachUserId:
          type: integer
          nullable: true
          description: If null, sends to all coaches in user's pod
        subject:
          type: string
        body:
          type: string
      required: [body]

    # ─── INVITE LINKS ─────────────────────────────────────────────────────────
...