# lib/db/src/schema/users.ts
import { pgTable, text, serial, timestamp, boolean, integer, jsonb, index } from "drizzle-orm/pg-core";
import { sql } from "drizzle-orm";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { householdsTable } from "./households";

export const userRoleEnum = ["super_admin", "coach", "parent", "student"] as const;
export type UserRole = (typeof userRoleEnum)[number];

export function isOperationalStaffRole(role: string | null | undefined): role is "super_admin" | "coach" {
  return role === "super_admin" || role === "coach";
}

export function isSuperAdminRole(role: string | null | undefined): role is "super_admin" {
  return role === "super_admin";
}

export const coachCertLevelEnum = ["1", "2", "3"] as const;
export type CoachCertLevel = (typeof coachCertLevelEnum)[number];

export const genderEnum = ["male", "female", "non_binary", "prefer_not_to_say"] as const;
export type Gender = (typeof genderEnum)[number];

export const seasonParticipationStatusEnum = ["active", "season_off", "pending"] as const;
export type SeasonParticipationStatus = (typeof seasonParticipationStatusEnum)[number];

export const usersTable = pgTable("users", {
  id: serial("id").primaryKey(),
  householdId: integer("household_id").references(() => householdsTable.id, { onDelete: "set null" }),
  firstName: text("first_name").notNull(),
  lastName: text("last_name").notNull(),
  email: text("email").notNull().unique(),
  phone: text("phone"),
  role: text("role", { enum: userRoleEnum }).notNull().default("parent"),
  podId: text("pod_id"),
  avatarUrl: text("avatar_url"),
  isActive: boolean("is_active").notNull().default(true),
  seasonParticipationStatus: text("season_participation_status", { enum: seasonParticipationStatusEnum }).notNull().default("active"),
  seasonParticipationSeasonId: integer("season_participation_season_id"),
  clerkUserId: text("clerk_user_id").unique(),
  gender: text("gender", { enum: genderEnum }),
  dateOfBirth: timestamp("date_of_birth", { withTimezone: true }),
  grade: integer("grade"),
  coachCertLevel: text("coach_cert_level", { enum: coachCertLevelEnum }),
  allergies: text("allergies"),
  medications: text("medications"),
  medicalNotes: text("medical_notes"),
  approved: boolean("approved").notNull().default(false),
  notificationsEnabled: boolean("notifications_enabled").notNull().default(true),
  emailNotifications: boolean("email_notifications").notNull().default(true),
  smsNotifications: boolean("sms_notifications").notNull().default(false),
  pushNotifications: boolean("push_notifications").notNull().default(true),
  defaultCarpoolSeats: integer("default_carpool_seats"),
  defaultCarpoolTrays: integer("default_carpool_trays"),
  notificationPreferencesLocked: boolean("notification_preferences_locked").notNull().default(false),
  notificationPreferences: jsonb("notification_preferences").$type<{
    practiceReminders: boolean;
    coachMessages: boolean;
    carpoolUpdates: boolean;
    eventReminders: boolean;
    rosterUpdates: boolean;
    boardReplies: boolean;
  }>().default(sql`'{"practiceReminders":true,"coachMessages":true,"carpoolUpdates":true,"eventReminders":true,"rosterUpdates":true,"boardReplies":true}'::jsonb`).$defaultFn(() => ({
    practiceReminders: true,
    coachMessages: true,
    carpoolUpdates: true,
    eventReminders: true,
    rosterUpdates: true,
    boardReplies: true,
  })),
  boardLastSeenAt: timestamp("board_last_seen_at", { withTimezone: true }),
  calendarToken: text("calendar_token").unique(),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
}, (t) => [
  index("users_household_id_idx").on(t.householdId),
  index("users_pod_id_idx").on(t.podId),
  index("users_role_idx").on(t.role),
]);

export const insertUserSchema = createInsertSchema(usersTable).omit({
  id: true,
  createdAt: true,
  updatedAt: true,
});

export type InsertUser = z.infer<typeof insertUserSchema>;
export type User = typeof usersTable.$inferSelect;